refactor: 💡 Migrate SQFormDatePickerWithCalendarInputOnly to TS
✅ Closes: #432
